Actualizacion. Extraccion de clases
