+++ README.md
@@ -5,3 +5,9 @@
 - Crea una branca del main per a afegir els teus canvis
 - Deixa la teva branca preparada per a que, en cas de fusionar-la amb la master, no afegeixi el teu fitxer .iml
 - No fusionis en cap cas la teva branca amb la master
+
+Hemos hecho extraccion de clases con las clases:
+-Customer
+-Human
+-Order
+-OrderLineItem
